Update contract manager
delegate some method to BundleService from ContractManager.
bring LogIndex method to ContractManager from ContractExecutor.
