+++ yggdrash-core/src/main/java/io/yggdrash/core/blockchain/osgi/ContractManager.java
@@ -6,6 +6,7 @@ import io.yggdrash.common.contract.BranchContract;
 import io.yggdrash.common.contract.ContractVersion;
 import io.yggdrash.core.blockchain.BranchId;
 import io.yggdrash.core.blockchain.Log;
+import io.yggdrash.core.blockchain.LogIndexer;
 import io.yggdrash.core.blockchain.SystemProperties;
 import io.yggdrash.core.blockchain.Transaction;
 import io.yggdrash.core.blockchain.genesis.GenesisBlock;
@@ -19,7 +20,6 @@ import io.yggdrash.core.store.LogStore;
 import org.apache.commons.io.IOUtils;
 import org.osgi.framework.Bundle;
 import org.osgi.framework.BundleException;
-import org.osgi.framework.Version;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -27,14 +27,9 @@ import java.io.File;
 import java.io.FileInputStream;
 import java.io.IOException;
 import java.io.InputStream;
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.Dictionary;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.jar.JarFile;
-import java.util.jar.Manifest;
 
 public class ContractManager {
     private static final Logger log = LoggerFactory.getLogger(ContractManager.class);
@@ -48,6 +43,7 @@ public class ContractManager {
     private final ContractExecutor contractExecutor;
 
     private final BundleService bundleService;
+    private final LogIndexer logIndexer;
 
     private final DefaultConfig defaultConfig;
     private final GenesisBlock genesis;
@@ -66,7 +62,8 @@ public class ContractManager {
 
         this.systemProperties = systemProperties;
 
-        this.contractExecutor = new ContractExecutor(contractStore, logStore);
+        this.logIndexer = new LogIndexer(logStore, contractStore.getTransactionReceiptStore());
+        this.contractExecutor = new ContractExecutor(contractStore, logIndexer);
 
         this.bundleService = bundleService;
         this.defaultConfig = defaultConfig;
@@ -75,15 +72,8 @@ public class ContractManager {
         this.serviceMap = new HashMap<>();
 
         initBootBundles();
+        initNodeContract();
 
-        VersioningContract service = new VersioningContract();
-        serviceMap.put(ContractConstants.VERSIONING_TRANSACTION, service);
-
-        try {
-            contractExecutor.injectNodeContract(service);
-        } catch (IllegalAccessException e) {
-            log.error(e.getMessage());
-        }
     }
 
     private void initBootBundles() {
@@ -127,7 +117,7 @@ public class ContractManager {
             }
 
             try {
-                start(bundle);
+                start(contractVersion);
             } catch (BundleException e) {
                 log.error("Bundle {} failed to start with {}", bundle.getSymbolicName(), e.getMessage());
                 continue;
@@ -138,8 +128,20 @@ public class ContractManager {
             try {
                 inject(bundle);
             } catch (IllegalAccessException e) {
-                e.printStackTrace();
+                log.warn(e.getMessage());
             }
+        }
+    }
+
+
+    private void initNodeContract() {
+        VersioningContract service = new VersioningContract();
+        serviceMap.put(ContractConstants.VERSIONING_TRANSACTION, service);
+
+        try {
+            contractExecutor.injectNodeContract(service);
+        } catch (IllegalAccessException e) {
+            log.error(e.getMessage());
         }
     }
 
@@ -171,43 +173,9 @@ public class ContractManager {
         return bundleService.getBundles();
     }
 
-    public Bundle install(ContractVersion contractVersion, File contractFile) throws IOException, BundleException {
-        Bundle bundle = bundleService.getBundle(contractVersion);
-
-        try (JarFile jarFile = new JarFile(contractFile)) {
-            if (bundle != null && isInstalledContract(jarFile, bundle)) {
-                log.debug("Already installed bundle {}", contractVersion);
-                return bundle;
-            }
-
-            if (verifyManifest(jarFile.getManifest())) {
-                log.debug("Installing  bundle {}", contractVersion);
-                return bundleService.install(contractVersion, contractFile);
-            }
-        }
-        return null;
-    }
-
     public Bundle install(ContractVersion contractVersion) throws IOException, BundleException {
-        File contractFile = new File(defaultConfig.getContractPath() + File.separator + contractVersion + ".jar");
-        return install(contractVersion, contractFile);
-    }
-
-    private boolean isInstalledContract(JarFile jarFile, Bundle bundle) throws IOException {
-        List<String> bundleKeys = Collections.list(bundle.getHeaders().keys());
-        Manifest m = jarFile.getManifest();
-
-        for (String key : bundleKeys) {
-            if (!m.getMainAttributes().getValue(key).equals(bundle.getHeaders().get(key))) {
-                return false;
-            }
-        }
-        return true;
-    }
-
-    public void registerServiceMap(ContractVersion contractVersion, Bundle bundle) {
-        Object service = bundleService.getBundleService(bundle);
-        this.serviceMap.put(contractVersion.toString(), service);
+        File contractFile = new File(contractFilePath(contractVersion));
+        return bundleService.install(contractVersion, contractFile);
     }
 
     public void uninstall(ContractVersion contractVersion) {
@@ -219,12 +187,12 @@ public class ContractManager {
     }
 
     public void start(ContractVersion contractVersion) throws BundleException {
-        Bundle bundle = bundleService.getBundle(contractVersion);
-        start(bundle);
+        bundleService.start(contractVersion);
     }
 
-    private void start(Bundle bundle) throws BundleException {
-        bundleService.start(bundle);
+    public void registerServiceMap(ContractVersion contractVersion, Bundle bundle) {
+        Object service = bundleService.getBundleService(bundle);
+        this.serviceMap.put(contractVersion.toString(), service);
     }
 
     public Long getStateSize() { // TODO for BranchController -> remove this
@@ -235,73 +203,24 @@ public class ContractManager {
         return contractExecutor;
     }
 
-    public String getContractPath() {
-        return contractPath;
-    }
-
     public Log getLog(long index) {
-        return contractExecutor.getLog(index);
+        return logIndexer.getLog(index);
     }
 
     public List<Log> getLogs(long start, long offset) {
-        return contractExecutor.getLogs(start, offset);
+        return logIndexer.getLogs(start, offset);
     }
 
     public long getCurLogIndex() {
-        return contractExecutor.getCurLogIndex();
-    }
-
-    private boolean verifyManifest(Manifest manifest) {
-        String manifestVersion = manifest.getMainAttributes().getValue("Bundle-ManifestVersion");
-        String bundleSymbolicName = manifest.getMainAttributes().getValue("Bundle-SymbolicName");
-        String bundleVersion = manifest.getMainAttributes().getValue("Bundle-Version");
-        return verifyManifest(manifestVersion, bundleSymbolicName, bundleVersion);
-    }
-
-    private boolean verifyManifest(String manifestVersion, String bundleSymbolicName, String bundleVersion) {
-        if (!"2".equals(manifestVersion)) {
-            log.error("Must set Bundle-ManifestVersion to 2");
-            return false;
-        }
-        if (bundleSymbolicName == null || "".equals(bundleSymbolicName)) {
-            log.error("Must set Bundle-SymbolicName");
-            return false;
-        }
-
-        if (bundleVersion == null || "".equals(bundleVersion)) {
-            log.error("Must set Bundle-Version");
-            return false;
-        }
-
-        return true;
+        return logIndexer.curIndex();
     }
 
     public List<ContractStatus> searchContracts() {
-        List<ContractStatus> result = new ArrayList<>();
-        Bundle[] bundleList = bundleService.getBundles();
-        for (Bundle bundle : bundleList) {
-            result.add(getContractStatus(bundle));
-        }
-        return result;
+        return bundleService.getContractList();
     }
 
-    private ContractStatus getContractStatus(Bundle bundle) {
-        Dictionary<String, String> header = bundle.getHeaders();
-        int serviceCnt = bundle.getRegisteredServices() == null ? 0 : bundle.getRegisteredServices().length;
 
-        Version v = bundle.getVersion();
-        return new ContractStatus(
-                bundle.getSymbolicName(),
-                String.format("%s.%s.%s", v.getMajor(), v.getMinor(), v.getMicro()),
-                header.get("Bundle-Vendor"),
-                header.get("Bundle-Description"),
-                bundle.getBundleId(),
-                bundle.getLocation(),
-                bundle.getState(),
-                serviceCnt
-        );
-    }
-
+    // Execute action
     public Object query(String contractVersion, String methodName, JsonObject params) {
         return contractExecutor.query(serviceMap, contractVersion, methodName, params);
     }
@@ -319,14 +238,11 @@ public class ContractManager {
         contractExecutor.commitBlockResult(result);
     }
 
-    public void close() {
-        contractStore.close();
-        logStore.close();
-    }
 
+    // file action.
     public boolean isContractFileExist(ContractVersion version) {
 
-        File contractDir = new File(this.getContractPath());
+        File contractDir = new File(contractPath);
         if (!contractDir.exists()) {
             contractDir.mkdirs();
             return false;
@@ -361,7 +277,9 @@ public class ContractManager {
         return this.contractPath + File.separator + contractVersion + ".jar";
     }
 
-    public Map<String, Object> getServiceMap() {
-        return serviceMap;
+    public void close() {
+        contractStore.close();
+        logStore.close();
     }
+
 }
